pyproject.toml
[project]
name = "stable_diffusion_backend"
version = "0.1.0"
description = "Stable Diffusion Backend"

# NOTE: Overwrites .python-version
# requires-python = ">=3.11,<3.13"

authors = [
    { name = "İlker SIĞIRCI", email = "sigirci.ilker@gmail.com" }
]
maintainers = [
    { name = "İlker SIĞIRCI", email = "sigirci.ilker@gmail.com" }
]
readme = "README.md"

keywords = [
    "template"
]

classifiers = [
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi[all]>=0.108.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic",
    "aiosqlite",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    # "httpx",
    "diffusers>=0.25.0",
    "torch>=2.1.2",
    "transformers",
    "accelerate",
    "omegaconf",
]

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit",
    "ruff==0.1.11",
    "mypy",
    "scalene~=1.5.21.2",
    "ipykernel",
    ## DOCS
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    # "mdx-include",
    # "mkdocs-markdownextradata-plugin",
    ## TEST
    "pytest",
    "pytest-cov",
    "pytest-rerunfailures",
    "pytest-xdist",
    "pytest-asyncio",
    # "pytest-timeout",
]

[[tool.rye.sources]]
name = "default"
url = "https://pypi.org/simple/"

[tool.pytest.ini_options]
minversion = "7.0.0"
addopts = [
    "--import-mode=importlib",
    "--color=yes",
    "--verbose",
#    "--durations=0",
#    "--strict-markers",
#    "--doctest-modules",
    # NOTE: cov is not working with pycharm debugger. Enable it later!
#    "--cov stable_diffusion_backend",
#    "--cov-report=xml",
#    "--cov-report html",
]

testpaths = ["tests"]

# TODO: Make them inherit from yaml file
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)-20s :: %(name)-8s :: %(levelname)-8s :: %(filename)s:%(lineno)d - (%(funcName)s) :: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff]
target-version = "py310"
# required-version = "0.1.11"
line-length = 88
src = ["src"]
select = [
    "B",  # flake8-bugbear
    "C",  # flake8-comprehensions
    "C9",  # mccabe
    # "D",  # flake8-docstrings
    "E",  # pycodestyle errors (default)
    "F",  # pyflakes (default)
    "I",  # isort
    "PD", # pandas-vet
    "PIE", # pie
    "PL",  # pylint
    "PTH", # pathlib
    "Q",  # flake8-quotes
    "RET", # return
    "RUF", # Enable all ruff-specific checks
    "SIM", # simplify
    "T20",  # (disallow print statements) keep debugging statements out of the codebase
    "W",  # pycodestyle warnings
]
#extend-select = []
respect-gitignore = true
# fix = true
unfixable = [
    "F401" # Unused imports.
]
ignore-init-module-imports = true
# ignore = []
extend-include = ["*.ipynb"]
extend-ignore = [
    "E501",  # Line too long, handled by ruff formatter
    "D107", # "Missing docstring in __init__",
]
extend-exclude = [
    "docs",
    "tests",
]

# Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]
# "path/to/file.py" = ["E402"]
# 'stable_diffusion_backend/__init__.py' = ['F405', 'F403']

# [tool.ruff.flake8-quotes]
# inline-quotes = "single"
# multiline-quotes = "double"
# docstring-quotes = "double"

[tool.ruff.format]
skip-magic-trailing-comma = false
# indent-style = "space"
# preview = true

[tool.ruff.flake8-import-conventions.extend-aliases]
some_alias = ""

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.isort]
combine-as-imports = true
# force-wrap-aliases = true
# known-first-party = ["SomeLibrary"]
# known-third-party = ["SomeLibrary"]

# mypy section
# Read more here: https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml-file
[tool.mypy]
# python_version = "3.8"
files = ["src"]
# exclude = ['^tests/']
strict = true
show_error_codes = true
# ignore_missing_imports = true

# TODO: Check whether this is working
[[tool.mypy.overrides]]
module = "tests.*"
ignore_missing_imports = true
check_untyped_defs = true

[tool.coverage.run]
source = ['src']
branch = true
context = '${CONTEXT}'

[tool.coverage.report]
precision = 2
exclude_lines = [
    'pragma: no cover',
    'raise NotImplementedError',
    'if TYPE_CHECKING:',
    '@overload',
]
# show_missing = True

# [tool.coverage.paths]
# source = [
#     'src/stable_diffusion_backend/',
# ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/stable_diffusion_backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
